fix: rename controllers to avoid naming conflicts with headless
