+++ Ucommerce.Sitefinity.UI/Properties/AssemblyInfo.cs
@@ -35,8 +35,8 @@ using UCommerce.Sitefinity.UI.App_Start;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("0.8.0.0")]
-[assembly: AssemblyFileVersion("4.0.1.0")]
+[assembly: AssemblyVersion("4.0.3.23013")]
+[assembly: AssemblyFileVersion("4.0.3.23013")]
 
 [assembly: ControllerContainer]
 [assembly: ResourcePackage]
